refactor: the delete logic
